+++ metaroom_xml_parser/include/simple_summary_parser.h
