+++ metaroom_xml_parser/include/simple_xml_parser.h
