+++ metaroom_xml_parser/src/simple_summary_parser.cpp
@@ -1,4 +1,4 @@
-#include "simpleSummaryParser.h"
+#include "simple_summary_parser.h"
 
 SimpleSummaryParser::SimpleSummaryParser(std::string xmlFile)
 {
